--- api/src/controllers/index.js
@@ -1,9 +1,0 @@
-import user from './UserController';
-import event from './EventController';
-import center from './CenterController';
-
-module.exports = {
-  user,
-  event,
-  center,
-};
